Updated load context example to actually show unloading.
